Gallery is finally here stable
Squashed some bugs in search page, removed more clutter from computer_course.html and
and GALLERY is there

Images used are used for educational purposes. We respect the IP rights of original creators.
If your creation/images are used in accident, you can raise a issue request to get it removed.

Thanks to @Manish and @Krishna